+++ dark/sam.py
@@ -1,3 +1,6 @@
+import six
+
+from itertools import chain
 from contextlib import contextmanager
 from collections import Counter, defaultdict
 
@@ -9,11 +12,15 @@ from dark.reads import Read, DNARead
 
 
 class UnequalReferenceLengthError(Exception):
-    "The reference sequences in a SAM/BAM file are not all of the same length."
+    "The references of interest in a SAM/BAM file are not of the same length."
 
 
 class UnknownReference(Exception):
     "Reference sequence not found in SAM/BAM file."
+
+
+class InvalidSAM(Exception):
+    "SAM/BAM file has unexpected/invalid content."
 
 
 # From https://samtools.github.io/hts-specs/SAMv1.pdf
@@ -21,81 +28,303 @@ _CONSUMES_QUERY = {CMATCH, CINS, CSOFT_CLIP, CEQUAL, CDIFF}
 _CONSUMES_REFERENCE = {CMATCH, CDEL, CREF_SKIP, CEQUAL, CDIFF}
 
 
+@contextmanager
+def samfile(filename):
+    """
+    A context manager to open and close a SAM/BAM file.
+
+    @param filename: A C{str} file name to open.
+    """
+    f = AlignmentFile(filename)
+    yield f
+    f.close()
+
+
+def samReferencesToStr(filenameOrSamfile, indent=0):
+    """
+    List SAM/BAM file reference names and lengths.
+
+    @param filenameOrSamfile: Either a C{str} SAM/BAM file name or an
+        instance of C{pysam.AlignmentFile}.
+    @param indent: An C{int} number of spaces to indent each line.
+    @return: A C{str} describing known reference names and their lengths.
+    """
+    indent = ' ' * indent
+
+    def _references(sam):
+        result = []
+        for i in range(sam.nreferences):
+            result.append('%s%s (length %d)' % (
+                indent, sam.get_reference_name(i), sam.lengths[i]))
+        return '\n'.join(result)
+
+    if isinstance(filenameOrSamfile, six.string_types):
+        with samfile(filenameOrSamfile) as sam:
+            return _references(sam)
+    else:
+        return _references(sam)
+
+
+class SAMFilter(object):
+    """
+    Filter a SAM/BAM file.
+
+    @param filename: The C{str} name of a BAM/SAM file to filter.
+    @param filterRead: A function that takes a C{dark.reads.Read) instance
+        and returns either C{None} or a C{Read} instance, according to
+        whether the passed read should be omitted or not. If C{None} is
+        passed, no filtering on reads (i.e., queries) is done.
+    @param referenceIds: Either C{None} or a set of C{str} reference ids
+        that should be kept (other references will be dropped).
+    @param storeQueryIds: If C{True}, query ids will be stored (in
+        C{self.queryIds}) as the SAM/BAM file is read.
+    @param dropUnmapped: If C{True}, unmapped matches will be excluded.
+    @param dropSecondary: If C{True}, secondary matches will be excluded.
+    @param dropSupplementary: If C{True}, supplementary matches will be
+        excluded.
+    @param dropDuplicates: If C{True}, matches flagged as optical or PCR
+        duplicates will be excluded.
+    @param keepQCFailures: If C{True}, reads that are considered quality
+        control failures will be included.
+    @param minScore: If not C{None}, alignments with score tag values less
+        than this value will not be output. If given, alignments that do not
+        have a score will not be output.
+    @param maxScore: If not C{None}, alignments with score tag values greater
+        than this value will not be output. If given, alignments that do not
+        have a score will not be output.
+    @param scoreTag: The alignment tag to extract for minScore and maxScore
+        comparisons.
+    """
+    def __init__(self, filename, filterRead=None, referenceIds=None,
+                 storeQueryIds=False, dropUnmapped=False,
+                 dropSecondary=False, dropSupplementary=False,
+                 dropDuplicates=False, keepQCFailures=False, minScore=None,
+                 maxScore=None, scoreTag='AS'):
+        self.filename = filename
+        self.filterRead = filterRead
+        self.referenceIds = referenceIds
+        self.dropUnmapped = dropUnmapped
+        self.dropSecondary = dropSecondary
+        self.dropSupplementary = dropSupplementary
+        self.dropDuplicates = dropDuplicates
+        self.keepQCFailures = keepQCFailures
+        self.storeQueryIds = storeQueryIds
+        self.minScore = minScore
+        self.maxScore = maxScore
+        self.scoreTag = scoreTag
+
+    @staticmethod
+    def addFilteringOptions(parser, samfileIsPositionalArg=False):
+        """
+        Add options to an argument parser for filtering SAM/BAM.
+
+        @param samfileIsPositionalArg: If C{True} the SAM/BAM file must
+            be given as the final argument on the command line (without
+            being preceded by --samfile).
+        @param parser: An C{argparse.ArgumentParser} instance.
+        """
+        parser.add_argument(
+            '%ssamfile' % ('' if samfileIsPositionalArg else '--'),
+            required=True,
+            help='The SAM/BAM file to filter.')
+
+        parser.add_argument(
+            '--referenceId', metavar='ID', nargs='+', action='append',
+            help=('A reference sequence id whose alignments should be kept '
+                  '(alignments against other references will be dropped). '
+                  'If omitted, alignments against all references will be '
+                  'kept. May be repeated.'))
+
+        parser.add_argument(
+            '--dropUnmapped', default=False, action='store_true',
+            help='If given, unmapped matches will not be output.')
+
+        parser.add_argument(
+            '--dropSecondary', default=False, action='store_true',
+            help='If given, secondary matches will not be output.')
+
+        parser.add_argument(
+            '--dropSupplementary', default=False, action='store_true',
+            help='If given, supplementary matches will not be output.')
+
+        parser.add_argument(
+            '--dropDuplicates', default=False, action='store_true',
+            help=('If given, matches flagged as optical or PCR duplicates '
+                  'will not be output.'))
+
+        parser.add_argument(
+            '--keepQCFailures', default=False, action='store_true',
+            help=('If given, reads that are considered quality control '
+                  'failures will be included in the output.'))
+
+        parser.add_argument(
+            '--minScore', type=float,
+            help=('If given, alignments with --scoreTag (default AS) values '
+                  'less than this value will not be output. If given, '
+                  'alignments that do not have a score will not be output.'))
+
+        parser.add_argument(
+            '--maxScore', type=float,
+            help=('If given, alignments with --scoreTag (default AS) values '
+                  'greater than this value will not be output. If given, '
+                  'alignments that do not have a score will not be output.'))
+
+        parser.add_argument(
+            '--scoreTag', default='AS',
+            help=('The alignment tag to extract for --minScore and --maxScore '
+                  'comparisons.'))
+
+    @classmethod
+    def parseFilteringOptions(cls, args, filterRead=None, storeQueryIds=False):
+        """
+        Parse command line options (added in C{addSAMFilteringOptions}.
+
+        @param args: The command line arguments, as returned by
+            C{argparse.parse_args}.
+        @param filterRead: A one-argument function that accepts a read
+            and returns C{None} if the read should be omitted in filtering
+            or else a C{Read} instance.
+        @param storeQueryIds: If C{True}, query ids will be stored as the
+            SAM/BAM file is read.
+        @return: A C{SAMFilter} instance.
+        """
+        referenceIds = (set(chain.from_iterable(args.referenceId))
+                        if args.referenceId else None)
+
+        return cls(
+            args.samfile,
+            filterRead=filterRead,
+            referenceIds=referenceIds,
+            storeQueryIds=storeQueryIds,
+            dropUnmapped=args.dropUnmapped,
+            dropSecondary=args.dropSecondary,
+            dropSupplementary=args.dropSupplementary,
+            dropDuplicates=args.dropDuplicates,
+            keepQCFailures=args.keepQCFailures,
+            minScore=args.minScore,
+            maxScore=args.maxScore)
+
+    def alignments(self):
+        """
+        Get alignments from the SAM/BAM file, subject to filtering.
+        """
+        referenceIds = self.referenceIds
+        dropUnmapped = self.dropUnmapped
+        dropSecondary = self.dropSecondary
+        dropSupplementary = self.dropSupplementary
+        dropDuplicates = self.dropDuplicates
+        keepQCFailures = self.keepQCFailures
+        storeQueryIds = self.storeQueryIds
+        filterRead = self.filterRead
+        minScore = self.minScore
+        maxScore = self.maxScore
+        scoreTag = self.scoreTag
+
+        if storeQueryIds:
+            self.queryIds = queryIds = set()
+
+        count = 0
+        with samfile(self.filename) as samAlignment:
+            for count, alignment in enumerate(samAlignment.fetch(), start=1):
+                if storeQueryIds:
+                    queryIds.add(alignment.query_name)
+
+                if minScore is not None or maxScore is not None:
+                    try:
+                        score = alignment.get_tag(scoreTag)
+                    except KeyError:
+                        continue
+                    else:
+                        if ((minScore is not None and score < minScore) or
+                                (maxScore is not None and score > maxScore)):
+                            continue
+
+                if ((filterRead is None or
+                     filterRead(Read(alignment.query_name,
+                                     alignment.query_sequence,
+                                     alignment.qual))) and
+                    not (
+                        (referenceIds and
+                         alignment.reference_name not in referenceIds) or
+                        (alignment.is_unmapped and dropUnmapped) or
+                        (alignment.is_secondary and dropSecondary) or
+                        (alignment.is_supplementary and dropSupplementary) or
+                        (alignment.is_duplicate and dropDuplicates) or
+                        (alignment.is_qcfail and not keepQCFailures))):
+                    yield alignment
+
+        self.alignmentCount = count
+
+    def referenceLengths(self):
+        """
+        Get the lengths of wanted references.
+
+        @raise UnknownReference: If a reference id is not present in the
+            SAM/BAM file.
+        @return: A C{dict} of C{str} reference id to C{int} length with a key
+            for each reference id in C{self.referenceIds} or for all references
+            if C{self.referenceIds} is C{None}.
+        """
+        result = {}
+        with samfile(self.filename) as sam:
+            if self.referenceIds:
+                for referenceId in self.referenceIds:
+                    tid = sam.get_tid(referenceId)
+                    if tid == -1:
+                        raise UnknownReference(
+                            'Reference %r is not present in the SAM/BAM file.'
+                            % referenceId)
+                    else:
+                        result[referenceId] = sam.lengths[tid]
+            else:
+                result = dict(zip(sam.references, sam.lengths))
+
+        return result
+
+
 class PaddedSAM(object):
     """
     Obtain aligned (padded) queries from a SAM/BAM file.
 
-    @param filename: The C{str} name of the SAM/BAM file.
+    @param samFilter: A C{SAMFilter} instance.
+    @raises UnequalReferenceLengthError: If C{referenceName} is C{None}
+        and the reference sequence lengths in the SAM/BAM file are not all
+        identical.
+    @raises UnknownReference: If C{referenceName} does not exist.
     """
-    def __init__(self, filename):
-        self.samfile = AlignmentFile(filename)
+    def __init__(self, samFilter):
+        referenceLengths = samFilter.referenceLengths()
+
+        if len(set(referenceLengths.values())) != 1:
+            raise UnequalReferenceLengthError(
+                'Your %d SAM/BAM file reference sequence '
+                'lengths (%s) are not all identical.' % (
+                    len(referenceLengths),
+                    ', '.join(
+                        '%s=%d' % (id_, referenceLengths[id_])
+                        for id_ in sorted(referenceLengths))))
+
+        # Get the length of any of the sequences (they are all identical).
+        self.referenceLength = referenceLengths.popitem()[1]
+        self.samFilter = samFilter
         # self.referenceInsertions will be keyed by query id (the query
         # that would cause a reference insertion). The values will be lists
         # of 2-tuples, with each 2-tuple containing an offset into the
         # reference sequence and the C{str} of nucleotides that would be
         # inserted starting at that offset.
         self.referenceInsertions = defaultdict(list)
-        self.alignmentCount = 0
 
-    def close(self):
-        """
-        Close the opened SAM/BAM file.
-        """
-        self.samfile.close()
-
-    def referencesToStr(self, indent=0):
-        """
-        List the reference names and their lengths.
-
-        @param indent: An C{int} number of spaces to indent each line.
-        @return: A C{str} describing known reference names and their lengths.
-        """
-        samfile = self.samfile
-        result = []
-        indent = ' ' * indent
-        for i in range(samfile.nreferences):
-            result.append('%s%s (length %d)' % (
-                indent, samfile.get_reference_name(i), samfile.lengths[i]))
-        return '\n'.join(result)
-
-    def queries(self, referenceName=None, minLength=0, rcSuffix='',
-                dropSecondary=False, dropSupplementary=False,
-                dropDuplicates=False, allowDuplicateIds=False,
-                keepQCFailures=False, rcNeeded=False, padChar='-',
-                queryInsertionChar='N', addAlignment=False,
-                storeQueryIds=False):
+    def queries(self, rcSuffix='', rcNeeded=False, padChar='-',
+                queryInsertionChar='N', allowDuplicateIds=False,
+                addAlignment=False):
         """
         Produce padded (with gaps) queries according to the CIGAR string and
         reference sequence length for each matching query sequence.
 
-        @param referenceName: The C{str} name of the reference sequence to
-            print alignments for. This is only needed if the SAM/BAM alignment
-            was made against multiple references *and* they are not all of the
-            same length. If there is only one reference sequence or if all
-            reference sequences are of the same length, there is no need to
-            provide a reference name (i.e., pass C{None}).
-        @param minLength: Ignore queries shorter than this C{int} value. Note
-            that this refers to the length of the query sequence once it has
-            been aligned to the reference. The alignment may introduce
-            C{queryInsertionChar} characters into the read, and these are
-            counted towards its length because the alignment is assuming the
-            query is missing a base at those locations.
         @param rcSuffix: A C{str} to add to the end of query names that are
             reverse complemented. This is added before the /1, /2, etc., that
             are added for duplicated ids (if there are duplicates and
             C{allowDuplicateIds} is C{False}.
-        @param dropSecondary: If C{True}, secondary matches will not be
-            yielded.
-        @param dropSupplementary: If C{True}, supplementary matches will not be
-            yielded.
-        @param dropDuplicates: If C{True}, matches flagged as optical or PCR
-            duplicates will not be yielded.
-        @param allowDuplicateIds: If C{True}, repeated query ids (due to
-            secondary or supplemental matches) will not have /1, /2, etc.
-            appended to their ids. So repeated ids may appear in the yielded
-            FASTA.
-        @param keepQCFailures: If C{True}, reads that are marked as quality
-            control failures will be included in the output.
         @param rcNeeded: If C{True}, queries that are flagged as matching when
             reverse complemented should have reverse complementing when
             preparing the output sequences. This must be used if the program
@@ -109,39 +338,22 @@ class PaddedSAM(object):
             is inserted as a 'missing' query character (i.e., a base that can
             be assumed to have been lost due to an error) whose existence is
             necessary for the match to continue.
+        @param allowDuplicateIds: If C{True}, repeated query ids (due to
+            secondary or supplemental matches) will not have /1, /2, etc.
+            appended to their ids. So repeated ids may appear in the yielded
+            FASTA.
         @param addAlignment: If C{True} the reads yielded by the returned
             generator will also have an C{alignment} attribute, being the
-            C{pysam.AlignedSegment} for each query.
-        @param storeQueryIds: If C{True}, query ids will be stored in a
-            C{set} attribute called C{queryIds}.
-        @raises UnequalReferenceLengthError: If C{referenceName} is C{None}
-            and the reference sequence lengths in the SAM/BAM file are not all
-            identical.
-        @raises UnknownReference: If C{referenceName} does not exist.
+            C{pysam.AlignedSegment} for the query.
+        @raises InvalidSAM: If a query has an empty SEQ field and either there
+            is no previous alignment or the alignment is not marked as
+            secondary or supplementary.
         @return: A generator that yields C{Read} instances that are padded
             with gap characters to align them to the length of the reference
             sequence. See C{addAlignment}, above, to yield reads with the
             corresponding C{pysam.AlignedSegment}.
         """
-        samfile = self.samfile
-
-        if referenceName:
-            referenceId = samfile.get_tid(referenceName)
-            if referenceId == -1:
-                raise UnknownReference(
-                    'Reference %r is not present in the SAM/BAM file.'
-                    % referenceName)
-            referenceLength = samfile.lengths[referenceId]
-        else:
-            # No reference given. All references must have the same length.
-            if len(set(samfile.lengths)) != 1:
-                raise UnequalReferenceLengthError(
-                    'Your SAM/BAM file has %d reference sequences, and their '
-                    'lengths (%s) are not all identical.' % (
-                        samfile.nreferences,
-                        ', '.join(map(str, sorted(samfile.lengths)))))
-            referenceId = None
-            referenceLength = samfile.lengths[0]
+        referenceLength = self.referenceLength
 
         # Hold the count for each id so we can add /1, /2 etc to duplicate
         # ids (unless --allowDuplicateIds was given).
@@ -149,23 +361,9 @@ class PaddedSAM(object):
 
         MATCH_OPERATIONS = {CMATCH, CEQUAL, CDIFF}
         lastQuery = None
-        if storeQueryIds:
-            self.queryIds = queryIds = set()
 
-        for lineNumber, alignment in enumerate(samfile.fetch(), start=1):
-            self.alignmentCount += 1
-
-            if storeQueryIds:
-                queryIds.add(alignment.query_name)
-
-            if (alignment.is_unmapped or
-                    (alignment.is_secondary and dropSecondary) or
-                    (alignment.is_supplementary and dropSupplementary) or
-                    (alignment.is_duplicate and dropDuplicates) or
-                    (alignment.is_qcfail and not keepQCFailures) or
-                    (referenceId is not None and
-                     alignment.reference_id != referenceId)):
-                continue
+        for lineNumber, alignment in enumerate(
+                self.samFilter.alignments(), start=1):
 
             query = alignment.query_sequence
 
@@ -176,13 +374,13 @@ class PaddedSAM(object):
             if query is None:
                 if alignment.is_secondary or alignment.is_supplementary:
                     if lastQuery is None:
-                        raise ValueError(
+                        raise InvalidSAM(
                             'Query line %d has an empty SEQ field, but no '
                             'previous alignment is present.' % lineNumber)
                     else:
                         query = lastQuery
                 else:
-                    raise ValueError(
+                    raise InvalidSAM(
                         'Query line %d has an empty SEQ field, but the '
                         'alignment is not marked as secondary or '
                         'supplementary.' % lineNumber)
@@ -197,8 +395,8 @@ class PaddedSAM(object):
                 if rcSuffix:
                     alignment.query_name += rcSuffix
 
-            # Adjust the query id if it's a duplicate and we're not
-            # allowing duplicates.
+            # Adjust the query id if it's a duplicate and we're not allowing
+            # duplicates.
             if allowDuplicateIds:
                 queryId = alignment.query_name
             else:
@@ -301,14 +499,9 @@ class PaddedSAM(object):
             assert queryIndex == len(query)
 
             # We cannot test we consumed the entire reference.  The CIGAR
-            # string applies to (i.e., exhausts) the query and is silent about
-            # the part of the reference that to the right of the aligned query.
-
-            # Check the length restriction now that we have (possibly) added
-            # queryInsertionChar characters to pad the query out to the length
-            # it requires to match the reference.
-            if len(alignedSequence) < minLength:
-                continue
+            # string applies to (and exhausts) the query but is silent
+            # about the part of the reference that lies to the right of the
+            # aligned query.
 
             # Put gap characters before and after the aligned sequence so that
             # it is offset properly and matches the length of the reference.
@@ -323,15 +516,3 @@ class PaddedSAM(object):
                 read.alignment = alignment
 
             yield read
-
-
-@contextmanager
-def samfile(filename):
-    """
-    A context manager to open and close a SAM/BAM file.
-
-    @param filename: A C{str} file name to open.
-    """
-    f = AlignmentFile(filename)
-    yield f
-    f.close()